refactor(ebpf): apply clippy lints and harden wire header parsing
- Replace unwrap-based header field reads with bounds-checked
  record.get(..) slices in EventEnvelope::decode_record
- Collapse redundant closures, then_some, repeat_n, struct-update
  syntax and dead branches flagged by clippy across agent, api,
  ebpf and testing crates
- Add test-agent-all Taskfile entry and build the host detrix binary
  explicitly for the Keycloak MCP bridge so it ignores the stale
  workspace target directory
- Run the Rust client fixture via its compiled binary, resolving the
  target directory through `cargo metadata`, and cover it with a unit
  test

+++ crates/detrix-ebpf/src/probe/ringbuf.rs
@@ -321,7 +321,7 @@ pub fn parse_ring_buffer_event_with_envelope(
                 // heap graph is traversed.  If the type name does not prove one
                 // of the contracts, retain the existing nested/blob decoder.
                 let special_value = if is_probably_rust_type(&var.type_name) {
-                    match crate::rust_layout::infer(&var.type_name, *byte_size as usize) {
+                    match crate::rust_layout::infer(&var.type_name, *byte_size) {
                         Some(crate::rust_layout::RustLayoutContract::NicheOption {
                             pointer_offset,
                             word_size,
